#23 - close bug, #34 - basic UI, #37 - 1 trillion tiles

+++ src/state/app-state.ts
@@ -6,7 +6,7 @@ export const AppState = {
 	mapProject: new Value<MapProject | null>(null),
 	mapSize: new Value<Size>({ width: 0, height: 0 }),
 	palette: new Value<Uint8Array | null>(null),
-	map: new Value<Uint8Array | null>(null),
+	map: new Value<Uint16Array | null>(null),
 	tiles: new Value<Tiles | null>(null),
 
 	wglMap: new Value<WglMap | null>(null),
